requirements.txt
click
curses-menu
gpiozero
keyboard
pycryptodome
pytest
luma.oled
spidev